pyproject.toml
[tool.poetry]
name = "presence-service"
version = "0.1.0"
description = ""
authors = ["nandan <nandanramdani608@gmail.com>"]
readme = "README.md"
packages = [{include = "presence_service"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
tortoise-orm = "^0.20.0"
fastapi-pagination = "^0.12.9"
aerich = "^0.7.2"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
